pyproject.toml
[tool.poetry]
name = "nettowel"
version = "0.2.0"
description = "Network Automation Collection"
authors = ["ubaumann <github@m.ubaumann.ch>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.4.0"
rich = "^11.2.0"
Jinja2 = {version = "^3.0.3", optional = true}
ttp = {version = "^0.8.4", optional = true}
textfsm = {version = "^1.1.2", optional = true}
napalm = {version = "^3.3.1", optional = true}
netmiko = {version = "^4", optional = true}
scrapli = {version = "^2022.1.30", optional = true}
nornir = {version = "^3.2.0", optional = true}
"ruamel.yaml" = "^0.17.21"
jinja2schema = {version = "^0.1.4", optional = true}
qrcode = "^7.3.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^7.0.1"
mypy = "^0.931"

[tool.poetry.extras]
full = ["Jinja2", "ttp", "textfsm", "napalm", "netmiko", "scrapli", "nornir"]
jinja = ["Jinja2", "jinja2schema"]
ttp = ["ttp"]
textfsm = ["textfsm"]
napalm = ["napalm"]
netmiko = ["netmiko"]
scrapli = ["scrapli"]
nornir = ["nornir"]

[tool.poetry.scripts]
nettowel = 'nettowel.cli.main:run'
nt = 'nettowel.cli.main:run'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.8"
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
warn_unused_configs = true
warn_unused_ignores = true
warn_return_any = true
warn_redundant_casts = true

[[tool.mypy.overrides]]
module = "nornir.core"
ignore_errors = true